fix uninitialized vars and make unsigned 64bit ints

+++ src/rabin.h
@@ -11,18 +11,18 @@
 #define MAXSIZE (32*1024)
 
 struct rabin_t {
-    uint8_t window[WINSIZE];
-    unsigned int wpos;
-    unsigned int count;
-    unsigned int pos;
-    unsigned int start;
-    uint64_t digest;
+  uint8_t window[WINSIZE];
+  uint64_t wpos;
+  uint64_t count;
+  uint64_t pos;
+  uint64_t start;
+  uint64_t digest;
 };
 
 struct chunk_t {
-    unsigned int start;
-    unsigned int length;
-    uint64_t cut_fingerprint;
+  uint64_t start;
+  uint64_t length;
+  uint64_t cut_fingerprint;
 };
 
 extern struct chunk_t last_chunk;
@@ -31,7 +31,7 @@ struct rabin_t *rabin_init(void);
 void rabin_reset(struct rabin_t *h);
 void rabin_slide(struct rabin_t *h, uint8_t b);
 void rabin_append(struct rabin_t *h, uint8_t b);
-int rabin_next_chunk(struct rabin_t *h, uint8_t *buf, unsigned int len);
+int rabin_next_chunk(struct rabin_t *h, uint8_t *buf, uint64_t len);
 struct chunk_t *rabin_finalize(struct rabin_t *h);
 
 #endif
